fix: rename figure and synthesis extraction
