+++ src/llm_synthesis/metrics/figure_extraction/base.py
