+++ src/llm_synthesis/metrics/text_extraction/base.py
